fix: rename route name from /signup to /apply-supplier
